Add files via upload
COVID-19 is a global issue and touched every corner of the world, impacted health, economy, social and political domains. Even after the similarity in problems, every country has different response and consequences.
With the good availability of the data, this dashboard has been created for each phase of the lockdowns and unlocks of India, from 31st December 2019 to 12th October 2020. For comparison, COVID-19 cases of the Asian countries is being considered here.